Switch downloads controller over to being an override
